# api/openapi.yaml
openapi: 3.0.3
info:
  title: sshc Local API
  version: 0.1.0
paths:
...
  /api/v1/vpn:
    get:
      operationId: getVPNOverview
      parameters:
        # false なら、経路の状態を docker から読むのを待たずに答える。まだ確かめて
        # いなければ checking を真にする。画面はプロファイルの一覧を先に見せ、経路の
        # 状態はあとから読み直して埋める。
        - name: waitForRoutes
          in: query
          required: false
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Configured VPN profiles, their sessions, and whether this machine can open routes
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VPNOverview" }
        "401": { $ref: "#/components/responses/Problem" }
  /api/v1/vpn/profiles:
    post:
      operationId: createVPNProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VPNProfileRequest" }
      responses:
        "200":
          description: Profile created together with its secrets. A secret left under the same name is replaced, never inherited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VPNOverview" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/vpn/profiles/{name}:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 48 } }
    put:
      operationId: updateVPNProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VPNProfileRequest" }
      responses:
        "200":
          description: Profile updated. Omitted or empty secret fields keep the stored values; secrets of another backend are dropped
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VPNOverview" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: deleteVPNProfile
      responses:
        "200":
          description: Profile, its secrets, and the bindings that named it were removed in one write. Requires an unlocked vault
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VPNOverview" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/vpn/profiles/{name}/rename:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 48 } }
    post:
      operationId: renameVPNProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VPNRenameRequest" }
      responses:
        "200":
          description: The profile, its secrets and the bindings that named it moved together
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VPNOverview" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/vpn/profiles/{name}/reveal:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 48 } }
    post:
      operationId: revealVPNSecrets
      parameters:
        # POST /api/v1/actions で kind を vpn_profile.reveal、target をプロファイルの名前に
        # して受け取ったトークン。
        - name: X-SSHC-Action
          in: header
          required: true
          schema: { type: string, minLength: 43, maxLength: 43 }
      responses:
        "200":
          description: The stored secrets of the profile's backend, for the edit form; never cached. Empty when the profile has none yet
          headers:
            Cache-Control:
              schema: { type: string, enum: [no-store] }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VPNSecrets" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/vpn/profiles/{name}/logs:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 48 } }
    get:
      operationId: getVPNLogs
      responses:
        "200":
          description: What the sshc engine recorded while preparing the route, followed by the container's recent output, with every stored secret masked
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VPNLogs" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/vpn/profiles/{name}/session:
    parameters:
      - { name: name, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 48 } }
    post:
      operationId: startVPNSession
      responses:
        "200":
          description: The route is up and its relay is listening
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VPNOverview" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: stopVPNSession
      responses:
        "200":
          description: The container was stopped and its relay socket removed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VPNOverview" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/vpn/bindings:
    put:
      operationId: setConnectionVPN
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VPNBindingRequest" }
      responses:
        "200":
          description: The connection now reaches its host through the named profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VPNOverview" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
...
components:
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        detail: { type: string }
        path: { type: string }
        line: { type: integer }
        column: { type: integer }
        diagnostics: { type: array, items: { $ref: "#/components/schemas/Diagnostic" } }
        conflict: { $ref: "#/components/schemas/ConflictReport" }
        blockers: { type: array, items: { type: string } }
        currentVersion: { type: integer, minimum: 0 }
        requiredVersion: { type: integer, minimum: 1 }
        # blockers は拒否理由を安定したコードで返す。
        # field・reason・limit は、入力のどの項目をどの理由で受け取れなかったかを
        # 決まった語で返す。field は保存形式の JSON のパス、limit は長さや数の上限。
        # VPN の経路や接続先へ届かなかった応答（vpn_session_failed、vpn_target_failed、
        # vpn_destination_invalid）では reason だけが入る。
        field: { type: string }
        reason: { type: string }
        limit: { type: integer, minimum: 1 }
        # 設定ファイル（OpenVPN の .ovpn）の中の誤りでは、断った指示の名前が入る。行番号は
        # line に入る。どちらも engine が決めた語と数だけで、設定ファイルの中身は載せない。
        directive: { type: string }
...
    Diagnostic:
      type: object
      additionalProperties: false
      required: [severity, code]
      properties:
        severity: { type: string }
        code: { type: string }
        path: { type: string }
        absolute: { type: string }
        external: { type: boolean }
        line: { type: integer }
        detail: { type: string }
...
    VPNOverview:
      type: object
      additionalProperties: false
      required: [available, profiles, checking]
      properties:
        # このマシンで経路を作れるか。作れない理由の語は unavailable に入る。
        available: { type: boolean }
        unavailable:
          type: string
          enum: [vpn_docker_missing, vpn_docker_not_running]
        # 使えないときの docker の生の文。訳さずに添えるだけにする。
        detail: { type: string }
        profiles:
          type: array
          maxItems: 64
          items: { $ref: "#/components/schemas/VPNProfileStatus" }
        # 経路の状態と、このマシンで経路を作れるかを、まだ確かめていない。
        # waitForRoutes=false で、確かめるのを待たずに答えたときだけ真になる。
        checking: { type: boolean }
    VPNProfileStatus:
      type: object
      additionalProperties: false
      required: [profile, running, relaySocket, connections, openConnections]
      properties:
        profile: { $ref: "#/components/schemas/VPNProfile" }
        running: { type: boolean }
        # engine の中継のソケットの場所。開いていなければ空になる。値があれば、
        # トンネルは用意できている。CLI はここへ繋ぎ、1行目に接続先（`host:port`）を
        # 送る。engine は1行の JSON で答えてから、その接続先へのバイト列を運ぶ。
        relaySocket: { type: string }
        # このプロファイルを通る接続の alias。
        connections:
          type: array
          items: { type: string }
        # コンテナの中のトンネルの様子。経路が無ければ省略する。
        tunnel: { $ref: "#/components/schemas/VPNTunnel" }
        # いま経路を用意している段階。用意していなければ省略する。初回は image
        # だけで分単位になることがあるので、待っている人へ何を待っているかを示す。
        phase:
          type: string
          enum: [image, container, tunnel, approval]
        # いまこの経路を通っている接続（Terminal、SFTP、CLI の中継）の数。経路を
        # 止めると、これらの接続も切れる。画面は切断の前にこれを確かめる。
        openConnections: { type: integer, minimum: 0 }
    VPNTunnel:
      type: object
      additionalProperties: false
      properties:
        interface: { type: string }
        # トンネル側でこの端末が名乗っているアドレス。L2TPでは相手から受け取る。
        address: { type: string }
        since: { type: string }
        backend: { type: string }
    VPNRenameRequest:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 48 }
    VPNLogs:
      type: object
      additionalProperties: false
      required: [lines]
      properties:
        # 直近の出力。保存済みの秘密とその16進表記は伏せてある。
        lines: { type: string, maxLength: 65536 }
    VPNProfileRequest:
      type: object
      additionalProperties: false
      required: [profile]
      properties:
        profile: { $ref: "#/components/schemas/VPNProfile" }
        # 省略すると、保存済みの秘密をそのまま残す。設定だけを直すときに、
        # 秘密を入れ直させない。
        secrets: { $ref: "#/components/schemas/VPNSecrets" }
    VPNSecrets:
      type: object
      additionalProperties: false
      properties:
        # WireGuard の設定ファイル（wg-quick の形）の中身。鍵（PrivateKey と PresharedKey）を
        # 含むので Vault に置く。engine は行ごとに確かめ、使えない項目は problem の line と
        # directive で名指しする。
        wireguardConfig: { type: string, minLength: 1, maxLength: 16384 }
        l2tpPassword: { type: string, minLength: 1, maxLength: 256 }
        ipsecPsk: { type: string, minLength: 1, maxLength: 256 }
        openconnectPassword: { type: string, minLength: 1, maxLength: 256 }
        openconnectTotpSecret: { type: string, minLength: 1, maxLength: 512 }
        # OpenVPN の設定ファイル（.ovpn）の中身。鍵を含むことが多いので Vault に置く。
        openvpnConfig: { type: string, minLength: 1, maxLength: 65536 }
        openvpnPassword: { type: string, minLength: 1, maxLength: 256 }
        ikev2Password: { type: string, minLength: 1, maxLength: 256 }
        ikev2Psk: { type: string, minLength: 1, maxLength: 256 }
    VPNBindingRequest:
      type: object
      additionalProperties: false
      required: [alias, profile]
      properties:
        alias: { type: string, minLength: 1, maxLength: 64 }
        # 空にすると紐付けを外す。
        profile: { type: string, maxLength: 48 }
    VPNProfile:
      type: object
      additionalProperties: false
      required: [name, backend]
      properties:
        name: { type: string, minLength: 1, maxLength: 48 }
        backend:
          type: string
          enum: [wireguard, l2tp_ipsec, openconnect, openvpn, ikev2]
        # 接続先を VPN の中で名前解決する DNS サーバー（IPv4）。接続先を
        # アドレスで書く接続にしか使わないなら要らない。
        dns:
          type: array
          maxItems: 3
          items: { type: string, minLength: 7, maxLength: 15 }
        wireguard: { $ref: "#/components/schemas/WireGuardProfile" }
        l2tp: { $ref: "#/components/schemas/L2TPProfile" }
        openconnect: { $ref: "#/components/schemas/OpenConnectProfile" }
        openvpn: { $ref: "#/components/schemas/OpenVPNProfile" }
        ikev2: { $ref: "#/components/schemas/IKEv2Profile" }
    OpenVPNProfile:
      type: object
      additionalProperties: false
      required: [servers]
      properties:
        # 設定ファイル（Vault にある）の remote のサーバーを、書かれた順に重複を除いて
        # 並べたもの。engine は保存するときに設定ファイルと合うことを確かめる。
        servers:
          type: array
          minItems: 1
          maxItems: 64
          items: { type: string, minLength: 1, maxLength: 320 }
        # auth-user-pass で送るユーザー名。空ならユーザー名とパスワードを送らない。
        username: { type: string, maxLength: 256 }
    OpenConnectProfile:
      type: object
      additionalProperties: false
      required: [server, username]
      properties:
        # VPN装置の名前またはアドレス。名前はコンテナの中で引く。
        server: { type: string, minLength: 1, maxLength: 320 }
        username: { type: string, minLength: 1, maxLength: 256 }
        # サーバーのプロトコル（anyconnect / nc / pulse / gp / f5 / fortinet / array）。
        # 空なら anyconnect。値は engine が確かめる。
        protocol: { type: string, maxLength: 32 }
        # 相手の証明書を固定する指紋（`sha256:...`）。公的な認証局の証明書を
        # 使う装置では空でよい。
        serverCertificate: { type: string, maxLength: 128 }
        # 二段目の質問への答え方（`approve` または `totp`）。空なら答えない。
        # 値は engine が確かめる。
        secondFactor: { type: string, maxLength: 16 }
        # secondFactor が approve のときに送る語。空なら push。
        approvalWord: { type: string, maxLength: 32 }
    IKEv2Profile:
      type: object
      additionalProperties: false
      required: [server, authentication, identity]
      properties:
        # VPNサーバーの名前またはアドレス。名前はコンテナの中で名前解決する。
        server: { type: string, minLength: 1, maxLength: 320 }
        # こちらを認証する方式（`eap-mschapv2` または `psk`）。値は engine が確かめる。
        authentication: { type: string, maxLength: 16 }
        # こちらの ID。EAP ではユーザー名。
        identity: { type: string, minLength: 1, maxLength: 256 }
        # サーバーの ID。空ならサーバーの名前を使う。
        serverIdentity: { type: string, maxLength: 256 }
        # サーバーの証明書を確かめる CA の証明書（PEM）。空なら公的な認証局で確かめる。
        # EAP のときだけ使う。
        caCertificate: { type: string, maxLength: 16384 }
        # サーバーと暗号スイートが合わないときだけ書く。L2TPProfile と同じ書き方。
        ike: { type: string, maxLength: 256 }
        esp: { type: string, maxLength: 256 }
    L2TPProfile:
      type: object
      additionalProperties: false
      required: [server, username]
      properties:
        # VPN装置の名前またはアドレス。名前はコンテナの中で引く。
        server: { type: string, minLength: 1, maxLength: 320 }
        username: { type: string, minLength: 1, maxLength: 256 }
        # 古い装置と暗号方式が合わないときだけ書く。
        ike: { type: string, maxLength: 256 }
        esp: { type: string, maxLength: 256 }
    WireGuardProfile:
      type: object
      additionalProperties: false
      required: [servers]
      properties:
        # 設定ファイル（Vault にある VPNSecrets.wireguardConfig）の Endpoint のサーバーを、
        # 書かれた順に重複を除いて並べたもの。engine は保存するときに設定ファイルと合うことを
        # 確かめる。
        servers:
          type: array
          minItems: 1
          maxItems: 64
          items: { type: string, minLength: 1, maxLength: 320 }
        # server、peerPublicKey、address は、v0.40.0 までの項目の形で保存したプロファイル
        # だけが持つ。engine は、これと Vault の秘密鍵から設定ファイルを組み立てて使い、
        # reveal はその設定ファイルを返す。保存し直すと消える。保存要求で送っても使わない。
        server: { type: string, minLength: 3, maxLength: 320 }
        peerPublicKey: { type: string, minLength: 44, maxLength: 44 }
        address: { type: string, minLength: 9, maxLength: 43 }
...
    DiffLine:
      type: object
      additionalProperties: false
      required: [op, text]
      properties:
        op: { type: string }
        text: { type: string }
        oldLine: { type: integer }
        newLine: { type: integer }
...
    ConflictReport:
      type: object
      additionalProperties: false
      required: [path, externalChange, localChange]
      properties:
        path: { type: string }
        baseDigest: { type: string }
        diskDigest: { type: string }
        externalChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
        localChange: { type: array, items: { $ref: "#/components/schemas/DiffLine" } }
...
  responses:
    Problem:
      description: Request rejected
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }